+++ src/ts/components/subtitlesettings/subtitlesettingsmanager.ts
@@ -9,6 +9,7 @@ interface SubtitleSettings {
   fontSize?: string;
   fontStyle?: string;
   characterEdge?: string;
+  characterEdgeColor?: string;
   backgroundColor?: string;
   backgroundOpacity?: string;
   windowColor?: string;
@@ -33,6 +34,7 @@ export class SubtitleSettingsManager {
     fontSize: new SubtitleSettingsProperty<string>(this),
     fontStyle: new SubtitleSettingsProperty<string>(this),
     characterEdge: new SubtitleSettingsProperty<string>(this),
+    characterEdgeColor: new SubtitleSettingsProperty<string>(this),
     backgroundColor: new SubtitleSettingsProperty<string>(this),
     backgroundOpacity: new SubtitleSettingsProperty<string>(this),
     windowColor: new SubtitleSettingsProperty<string>(this),
@@ -73,6 +75,10 @@ export class SubtitleSettingsManager {
 
   public get characterEdge(): SubtitleSettingsProperty<string> {
     return this._properties.characterEdge;
+  }
+
+  public get characterEdgeColor(): SubtitleSettingsProperty<string> {
+    return this._properties.characterEdgeColor;
   }
 
   public get backgroundColor(): SubtitleSettingsProperty<string> {